requirements.txt
rich
rstatus
rich
requests
dnspython
subprocess